Introduce `Console::Format` for format handling.
